move build_root_params() into class function of RootParamsGroup
